# Support creating smaller eStargz images (`--estargz-external-toc` and `--estargz-min-chunk-size`)