… docs/ctr-remote.md
… docs/estargz.md
… docs/images/estargz-inneroffset.png
… docs/smaller-estargz.md
… estargz/build.go
… estargz/build_test.go
… estargz/estargz.go
… estargz/externaltoc/externaltoc.go
… estargz/externaltoc/externaltoc_test.go
… estargz/gzip_test.go
… estargz/testutil.go
… estargz/types.go
… estargz/zstdchunked/zstdchunked_test.go
… fs/fs.go
… fs/layer/layer.go
… fs/layer/testutil.go
… fs/reader/reader.go
… fs/reader/testutil.go
… metadata/memory/reader.go
… metadata/metadata.go
… metadata/testutil/testutil.go
… nativeconverter/estargz/externaltoc/converter.go
… nativeconverter/estargz/externaltoc/converter_test.go
… nativeconverter/estargz/externaltoc/fetcher.go
… script/demo/config.stargz.toml
… script/integration/containerd/entrypoint.sh
… service/service.go
… store/manager.go
… store/refs.go
… util/containerdutil/manifest.go